feat(#570): adding uiFormikValues to initialValues object (#608)
When the fields are not in schema formik doesn't register the field in initial values. Sometimes, there are data points (previously used in UI schema) that need to be used for the sole purpose of showing and hiding different fields. When these fields are not registered, formik doesn't hook into any validation functions for it. This needs to be added so formik registers it as an initial value.

Bumping version to 1.5.0 for this new feature.